requirements.txt
fastapi
python-jose
passlib
pymongo
python-multipart

# autopep8==1.6.0
# uvicorn[standard]==0.18.3
# gunicorn==20.1.0
# pydantic[dotenv]
# pydantic[email]
# python-jose[cryptography]
# passlib[bcrypt]